[Viz] Extract FDShapeArtist mixin with the shared shape draw hooks
